Successfully move one "user" endpoint into new package
