+++ crates/component_subsystem/users/Cargo.toml
@@ -16,6 +16,8 @@ path = "src/lib.rs"
 
 # Internal
 container_common = { path = "../../lib/container_common" }
+database_queries = { path = "../../lib/database_queries" }
+http_server_common = { path = "../../lib/http_server_common" }
 
 # External
 anyhow = "1.0.40"
@@ -40,3 +42,9 @@ serde = "1.0.125"
 serde_derive = "1.0.125"
 serde_json = "1.0.64"
 toml = "0.5.8"
+
+# MySql Database
+# "offline" supports typesafe builds without a live database
+# "rustls" doesn't appear to work with IP addresses, per:
+# https://github.com/launchbadge/sqlx/issues/1095#issuecomment-798050828
+sqlx = { version = "0.5", features = [ "mysql", "runtime-actix-native-tls", "offline", "chrono" ] }
