+++ crates/component_subsystem/users/src/endpoints/mod.rs
@@ -1,0 +1,1 @@
+pub mod session_info;
