+++ crates/component_subsystem/users/src/lib.rs
@@ -1,1 +1,4 @@
+#[macro_use] extern crate serde_derive;
+
+pub mod endpoints;
 pub mod utils;
