+++ crates/component_subsystem/users/src/utils/mod.rs
@@ -1,1 +1,2 @@
+pub mod session_checker;
 pub mod session_cookie_manager;
