+++ crates/lib/database_queries/src/helpers/boolean_converters.rs
@@ -15,6 +15,16 @@ pub fn nullable_i8_to_bool(value: Option<i8>, default_value: bool) -> bool {
       .unwrap_or(default_value)
 }
 
+pub fn nullable_i8_to_bool_default_false(value: Option<i8>) -> bool {
+  match value {
+    None => false,
+    Some(v) => match v {
+      0 => false,
+      _ => true,
+    }
+  }
+}
+
 /// Bool conversion, but retain nulls.
 pub fn nullable_i8_to_optional_bool(value: Option<i8>) -> Option<bool> {
   value.map(|v| i8_to_bool(v))
@@ -22,11 +32,7 @@ pub fn nullable_i8_to_optional_bool(value: Option<i8>) -> Option<bool> {
 
 #[cfg(test)]
 mod tests {
-  use crate::helpers::boolean_converters::{
-    i8_to_bool,
-    nullable_i8_to_bool,
-    nullable_i8_to_optional_bool
-  };
+  use crate::helpers::boolean_converters::{i8_to_bool, nullable_i8_to_bool, nullable_i8_to_bool_default_false, nullable_i8_to_optional_bool};
 
   #[test]
   fn test_i8_to_bool() {
@@ -50,6 +56,16 @@ mod tests {
     assert_eq!(nullable_i8_to_bool(Some(100), true), true);
     assert_eq!(nullable_i8_to_bool(Some(-100), false), true);
     assert_eq!(nullable_i8_to_bool(Some(-100), true), true);
+  }
+
+  #[test]
+  fn test_nullable_i8_to_bool_default_false() {
+    assert_eq!(nullable_i8_to_bool_default_false(None), false);
+    assert_eq!(nullable_i8_to_bool_default_false(Some(0)), false);
+
+    assert_eq!(nullable_i8_to_bool_default_false(Some(1)), true);
+    assert_eq!(nullable_i8_to_bool_default_false(Some(100)), true);
+    assert_eq!(nullable_i8_to_bool_default_false(Some(-100)), true);
   }
 
   #[test]
