+++ crates/lib/database_queries/src/queries/users/user_sessions/mod.rs
@@ -1,1 +1,3 @@
 pub mod create_session;
+pub mod get_session_by_token;
+pub mod get_session_by_token_light;
