+++ crates/service/storyteller_web/src/http_server/endpoints/tts/get_tts_result.rs
@@ -10,7 +10,6 @@ use chrono::{DateTime, Utc};
 use crate::AnyhowResult;
 use crate::http_server::web_utils::ip_address::get_request_ip;
 use crate::http_server::web_utils::response_error_helpers::to_simple_json_error;
-use crate::http_server::web_utils::session_checker::SessionRecord;
 use crate::server_state::ServerState;
 use crate::validations::passwords::validate_passwords;
 use crate::validations::username::validate_username;
