+++ crates/service/storyteller_web/src/http_server/web_utils/mod.rs
@@ -5,4 +5,3 @@ pub mod redis_rate_limiter;
 pub mod response_error_helpers;
 pub mod response_success_helpers;
 pub mod serialize_as_json_error;
-pub mod session_checker;
