+++ crates/service/storyteller_web/src/main.rs
@@ -86,7 +86,6 @@ use crate::http_server::endpoints::users::list_user_w2l_inference_results::list_
 use crate::http_server::endpoints::users::list_user_w2l_templates::list_user_w2l_templates_handler;
 use crate::http_server::endpoints::users::login::login_handler;
 use crate::http_server::endpoints::users::logout::logout_handler;
-use crate::http_server::endpoints::users::session_info::session_info_handler;
 use crate::http_server::endpoints::w2l::delete_w2l_result::delete_w2l_inference_result_handler;
 use crate::http_server::endpoints::w2l::delete_w2l_template::delete_w2l_template_handler;
 use crate::http_server::endpoints::w2l::edit_w2l_result::edit_w2l_inference_result_handler;
@@ -103,7 +102,6 @@ use crate::http_server::endpoints::w2l::list_w2l_templates::list_w2l_templates_h
 use crate::http_server::endpoints::w2l::set_w2l_template_mod_approval::set_w2l_template_mod_approval_handler;
 use crate::http_server::middleware::ip_filter_middleware::IpFilter;
 use crate::http_server::web_utils::redis_rate_limiter::RedisRateLimiter;
-use crate::http_server::web_utils::session_checker::SessionChecker;
 use crate::routes::add_routes;
 use crate::server_state::{ServerState, EnvConfig, TwitchOauthSecrets, TwitchOauth, RedisRateLimiters, InMemoryCaches};
 use crate::threads::db_health_checker_thread::db_health_check_status::HealthCheckStatus;
@@ -128,6 +126,7 @@ use std::time::Duration;
 use storage_buckets_common::bucket_client::BucketClient;
 use tokio::runtime::Runtime;
 use twitch_common::twitch_secrets::TwitchSecrets;
+use users_component::utils::session_checker::SessionChecker;
 use users_component::utils::session_cookie_manager::SessionCookieManager;
 
 // TODO TODO TODO TODO
@@ -436,6 +435,8 @@ pub async fn serve(server_state: ServerState) -> AnyhowResult<()>
     let build_sha = std::fs::read_to_string("/GIT_SHA").unwrap_or(String::from("unknown"));
 
     let app = App::new()
+      .app_data(web::Data::new(server_state_arc.mysql_pool.clone()))
+      .app_data(web::Data::new(server_state_arc.session_checker.clone()))
       .app_data(server_state_arc.clone())
       .wrap(build_common_cors_config())
       .wrap(DefaultHeaders::new()
