+++ crates/service/storyteller_web/src/server_state.rs
@@ -2,7 +2,6 @@ use crate::StaticApiTokenSet;
 use crate::http_server::endpoints::categories::list_tts_categories::DisplayCategory;
 use crate::http_server::endpoints::tts::list_tts_models::TtsModelRecordForResponse;
 use crate::http_server::web_utils::redis_rate_limiter::RedisRateLimiter;
-use crate::http_server::web_utils::session_checker::SessionChecker;
 use crate::threads::db_health_checker_thread::db_health_check_status::HealthCheckStatus;
 use crate::threads::ip_banlist_set::IpBanlistSet;
 use crate::util::encrypted_sort_id::SortKeyCrypto;
@@ -14,6 +13,7 @@ use memory_caching::single_item_ttl_cache::SingleItemTtlCache;
 use r2d2_redis::{r2d2, RedisConnectionManager};
 use sqlx::MySqlPool;
 use storage_buckets_common::bucket_client::BucketClient;
+use users_component::utils::session_checker::SessionChecker;
 use users_component::utils::session_cookie_manager::SessionCookieManager;
 
 /// State that is injected into every endpoint.
